+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Added license file, MIT, same as `package.json` already specifies
+- Added `messageMatched` events when subscribers on `Storage` match messages
 
 ## [0.0.1] - 2019-04-10
 ### Added